+++ game/server/ai_hint.h
@@ -151,8 +151,8 @@ public:
 	bool		InIncludedZone( const Vector &testPosition ) const;
 	bool		InExcludedZone( const Vector &testPosition ) const;
 
-	int			NumHintTypes() const;
-	int			GetHintType( int idx ) const;
+	intp		NumHintTypes() const;
+	int			GetHintType( intp idx ) const;
 
 private:
 
@@ -226,9 +226,9 @@ public:
 
 	// Purpose: Finds a random suitable hint within the requested radious of the npc
 	static CAI_Hint		*FindHintRandom( CAI_BaseNPC *pNPC, const Vector &position, const CHintCriteria &hintCriteria );
-	static int			FindAllHints( CAI_BaseNPC *pNPC, const Vector &position, const CHintCriteria &hintCriteria, CUtlVector<CAI_Hint *> *pResult );
-	static int			FindAllHints( const Vector &position, const CHintCriteria &hintCriteria, CUtlVector<CAI_Hint *> *pResult )	{ return FindAllHints( NULL, position, hintCriteria, pResult ); }
-	static int			FindAllHints( CAI_BaseNPC *pNPC, const CHintCriteria &hintCriteria, CUtlVector<CAI_Hint *> *pResult )		{ return FindAllHints( pNPC, pNPC->GetAbsOrigin(), hintCriteria, pResult ); }
+	static intp			FindAllHints( CAI_BaseNPC *pNPC, const Vector &position, const CHintCriteria &hintCriteria, CUtlVector<CAI_Hint *> *pResult );
+	static intp			FindAllHints( const Vector &position, const CHintCriteria &hintCriteria, CUtlVector<CAI_Hint *> *pResult )	{ return FindAllHints( NULL, position, hintCriteria, pResult ); }
+	static intp			FindAllHints( CAI_BaseNPC *pNPC, const CHintCriteria &hintCriteria, CUtlVector<CAI_Hint *> *pResult )		{ return FindAllHints( pNPC, pNPC->GetAbsOrigin(), hintCriteria, pResult ); }
 	static int			GetFlags( const char *token );
 
 	static CAI_Hint		*GetFirstHint( AIHintIter_t *pIter );
