+++ game/server/hierarchy.h
@@ -19,8 +19,8 @@ void		UnlinkFromParent( CBaseEntity *pRemove );
 void		TransferChildren( CBaseEntity *pOldParent, CBaseEntity *pNewParent );
 void		LinkChild( CBaseEntity *pParent, CBaseEntity *pChild );
 void		UnlinkAllChildren( CBaseEntity *pParent );
-int			GetAllChildren( CBaseEntity *pParent, CUtlVector<CBaseEntity *> &list );
+intp		GetAllChildren( CBaseEntity *pParent, CUtlVector<CBaseEntity *> &list );
 bool		EntityIsParentOf( CBaseEntity *pParent, CBaseEntity *pEntity );
-int			GetAllInHierarchy( CBaseEntity *pParent, CUtlVector<CBaseEntity *> &list );
+intp		GetAllInHierarchy( CBaseEntity *pParent, CUtlVector<CBaseEntity *> &list );
 
 #endif // HIERARCHY_H
